requirements.txt
django==2.2.*
factory-boy==2.12.0
social-auth-app-django
Pillow
django-stdimage
requests
django-widget-tweaks
django-ipware==2.1.0
djangorestframework==3.10.2
